third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/grid-block-start-expected.txt

This is a testharness.js-based test.
[FAIL] grid > item 1
  assert_equals: \n<item class="locked-location" data-expected-margin-top="0"></item>\nmargin-top expected "0" but got "25"
[FAIL] grid > item 2
  assert_equals: \n<item data-expected-margin-top="0"></item>\nmargin-top expected "0" but got "-30"
Harness: the test ran to completion.
